% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{em_ito}
\alias{em_ito}
\title{Solve a constant-coefficient SDE for Ito diffusion}
\usage{
em_ito(x0, t, drift, volat, n)
}
\arguments{
\item{x0}{initial value}

\item{t}{the time period to solver over}

\item{drift}{the drift-rate}

\item{volat}{the volatility coefficient}

\item{n}{number of variates in time-discretization}
}
\value{
numeric vector
}
\description{
{Solve the SDE for a constant-coefficient Ito diffusion.}
}
