% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{jump_sde_em}
\alias{jump_sde_em}
\title{Simulate a sample path of an exponential Ito-Levy process with the Euler-Maruyama scheme}
\usage{
jump_sde_em(spot, t, mu, volat, jumps = NULL, n = 1000, M = 20000)
}
\arguments{
\item{spot}{the initial price of the share-price process.}

\item{t}{the terminal time to simulate until}

\item{mu}{the infinitesimal drift coefficient function of \code{(t, x)}}

\item{volat}{the infinitesimal volatility coefficient function of \code{(t, x)}}

\item{jumps}{a list of objects defining the jump-size distribution, can be NULL for purely continuous models}

\item{n}{number of sub-intervals in time-grid}

\item{M}{number of samples to use in Monte-Carlo estimate of \code{eta} for the jump-diffusion drift correction.}
}
\value{
data.frame
}
\description{
Simulate a sample path of an exponential Ito-Levy process with the Euler-Maruyama scheme
}
\details{
{ while \code{jumps} should contain
\itemize{
\item \code{lambda} the mean-rate of jumps function as a function of \code{(t, x)}.
\item \code{distr} the name of the distribution of the jump-sizes: "norm", "unif", "kou"
\item \code{param} named list of parameters for the distribution matching the input in \code{rdistr} for a given "distr".
}}
}
