% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanJumpSize.R
\name{meanJumpSize}
\alias{meanJumpSize}
\title{Compute mean jump size given list of jump-dynamics variables}
\usage{
meanJumpSize(jump_specs, samples = 10000)
}
\arguments{
\item{jump_specs}{list of variables defining jump dynamics, see details}

\item{samples}{number of variates to use in Monte-Carlo estimation of mean jump-size}
}
\value{
numeric
}
\description{
Compute mean jump size given list of jump-dynamics variables
}
\details{
the list \code{jump_specs} should contain
\itemize{
\item \code{distr} the jump-size distribution name: "kou", "norm" or "unif",
\item \code{param} the list of parameters for the above distribution}
}
