% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePathWrappers.R
\name{samplePathItoLevy}
\alias{samplePathItoLevy}
\title{Simulate a sample-path of an exponential Ito-Levy process by solving a SDE with jumps}
\usage{
samplePathItoLevy(region, dynamics, jumps, n)
}
\arguments{
\item{region}{vector of initial price and time-horizon}

\item{dynamics}{list of drift and volatility coefficient functions}

\item{jumps}{the jump-specification list}

\item{n}{number of variates to discretize time by}
}
\value{
data.frame of time and price
}
\description{
{Solve an SDE with jumps via the Euler-Maruyama discretization and hence
simulate a sample-path of the stochastic process.}
}
\details{
{ The SDE solved is for the log-dynamics and
then the process is exponentiated and returned. Therefore the drift and volatility coefficents passed
must be those figuring in the SDE (without jumps for the sake of convenience, here)
\eqn{dS_t=\mu(t, S_t) S_t dt+\sigma(t, S_t) S_t dB_t}.

Further, the argument \code{jumps} must be a named list containing
\itemize{
\item \code{lambda} the mean-rate of jumps function as a function of \code{(t, x)}.
\item \code{distr} the name of the distribution of the (log-)jump-sizes
\item \code{param} a named list of the parameters of the distribution matching
the names and order of the arguments in e.g. \code{runif}, etc.}
}
}
