% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePathWrappers.R
\name{samplePathSystem}
\alias{samplePathSystem}
\title{Solve arbitrary first order SDE systems}
\usage{
samplePathSystem(f, g, IC, t0 = 0, tn = 1, independent = FALSE, n = 1000)
}
\arguments{
\item{f}{a list of drift functions with as many variables as IC, and in the same order with the names}

\item{g}{a list of volatility functions with as many variables as IC, and in the same order with the names}

\item{IC}{a list of variables with the same names as the non-time input as f, in the same order with initial values}

\item{t0}{initial time}

\item{tn}{ending time}

\item{independent}{boolean to simulate processes driven by one Brownian motion (false) or as many independent Brownian motions as there are state variables}

\item{n}{number of sub-intervals in time-grid}
}
\value{
data.frame
}
\description{
{A general Euler scheme implementation for arbitrary first-order SDE systems of finite dimension.}
}
\details{
{The list of functions must have syntax \code{f(t,x,y,z)} for each element matching the elements of \code{IC} as \code{list(x = y0, y = y0, z = z0)} for example.}
}
