% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePathWrappers.R
\name{samplePathTwoStateSystem}
\alias{samplePathTwoStateSystem}
\title{Solve arbitrary first order SDE systems}
\usage{
samplePathTwoStateSystem(
  drifts,
  diffusions,
  IC,
  t0 = 0,
  tn = 1,
  rho = 0,
  n = 1000
)
}
\arguments{
\item{drifts}{a list of drift functions with as many variables as IC, and in the same order with the names}

\item{diffusions}{a list of volatility functions with as many variables as IC, and in the same order with the names}

\item{IC}{a list of variables with the same names as the non-time input as drifts, in the same order with initial values}

\item{t0}{initial time}

\item{tn}{ending time}

\item{rho}{the correlation factor}

\item{n}{number of sub-intervals in time-grid}
}
\value{
data.frame
}
\description{
{A general Euler scheme implementation for arbitrary first-order SDE systems with two state variables.}
}
\details{
{The list of functions must have syntax \code{h(t,x,y)} for each element matching the elements of \code{IC} as \code{list(x = y0, y = y0)} for example.}
}
