% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{sde}
\alias{sde}
\title{Solve a one-dimensional SDE for arbitrary coefficient functions under three different numerical schemes}
\usage{
sde(x0, tt, dynamics, n = 1000, method = "milstein", plotG = TRUE)
}
\arguments{
\item{x0}{initial point of the process}

\item{tt}{the time horizon to solve over}

\item{dynamics}{list of functions defining dynamics of the stochastic process, see details}

\item{n}{number of time sub-intervals in the partition}

\item{method}{the numerical integration method to use: "em", "rk2", or "milstein", see details}

\item{plotG}{whether to plot the sample path in the function body}
}
\value{
numeric/data.frame
}
\description{
{Generate sample paths of processes solving one-dimensional SDEs with near arbitrary
coefficient functions. Three schemes are available: Euler-Maruyama, RK2 and Milstein's method.}
}
\details{
{For \code{method = "milstein"} one must pass single-variable coefficient functions and
one can optionally pass the derivative of the volatility function if it is known, otherwise use a numeric
approximation, as one of the two is required in the Milstein scheme. For \code{method = "em"} or "rk2", the
coefficient functions ought to be functions of \code{(t, x)} but wrappers are made if this is not done.

The \code{dynamics} list must be a list of either 2 or 3 functions defining the infinitesimal drift
function and the infinitesimal volatility function and optionally its derivative for the Milstein method.}
}
