% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{sde_em}
\alias{sde_em}
\title{Euler-Maruyama solver for Ito SDEs}
\usage{
sde_em(x0, t, drift, volat, n = 50)
}
\arguments{
\item{x0}{initial value of the process}

\item{t}{length to simulate over}

\item{drift}{the drift coefficient function of space-time}

\item{volat}{the volatility coefficient function of space-time}

\item{n}{number of sub-intervals in time-discretization}
}
\value{
data.frame
}
\description{
{The stochastic Euler-Maruyama solver for SDEs. See wikipedia for details}
}
