% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{sde_gbm}
\alias{sde_gbm}
\title{Simulate a geometric Brownian motion}
\usage{
sde_gbm(x0, tt, param, n = 1000, method = "em")
}
\arguments{
\item{x0}{initial value}

\item{tt}{time horizon}

\item{param}{vector of mean drift and volatility}

\item{n}{number of time sub-intervals}

\item{method}{"em" or "rk2"}
}
\value{
data.frame numeric
}
\description{
{Simulate a geometric Brownian motion sample path using
a numerical stochastic integrator.}
}
