% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{sde_heston}
\alias{sde_heston}
\title{Simulate a Heston stochastic volatility model}
\usage{
sde_heston(s0, v0, tt, param, n = 1000)
}
\arguments{
\item{s0}{initial price}

\item{v0}{initial volatility}

\item{tt}{time horizon}

\item{param}{vector of mean drift, correlation, and Heston parameters}

\item{n}{number of time sub-intervals}
}
\value{
data.frame numeric
}
\description{
{Simulate a Heston stochastic volatility model sample path using
a numerical stochastic integrator.}
}
