% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{sde_kou}
\alias{sde_kou}
\title{Simulate a Kou jump diffusion}
\usage{
sde_kou(x0, tt, param, n = 1000)
}
\arguments{
\item{x0}{initial value}

\item{tt}{time horizon}

\item{param}{vector of mean drift, volatility,
mean rate of jumps, probability of jump up, mean jump up size, and mean jump down size}

\item{n}{number of time sub-intervals}
}
\value{
data.frame numeric
}
\description{
{Simulate a Kou jump diffusion sample path using
a numerical stochastic integrator.}
}
