% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EulerMaruyama.R
\name{sde_levy}
\alias{sde_levy}
\title{Simulate a sample path of an exponential Ito-Levy process with the Euler-Maruyama scheme}
\usage{
sde_levy(spot, tt, drift, diffusion, lambda, jumps, n = 1000, M = 20000)
}
\arguments{
\item{spot}{the initial price of the share-price process.}

\item{tt}{the terminal time to simulate until}

\item{drift}{the infinitesimal drift coefficient function of \code{(t, x)}}

\item{diffusion}{the infinitesimal volatility coefficient function of \code{(t, x)}}

\item{lambda}{the mean jump-rate function of \code{(t, x)}}

\item{jumps}{a list of objects defining the jump-size distribution}

\item{n}{number of sub-intervals in time-grid}

\item{M}{number of samples to use in Monte-Carlo estimate of \code{eta} for the jump-diffusion drift correction.}
}
\value{
data.frame
}
\description{
Simulate a sample path of an exponential Ito-Levy process with the Euler-Maruyama scheme
}
\details{
{ while \code{jumps} should contain
\itemize{
\item \code{distr} the name of the distribution of the jump-sizes: "norm", "unif", "kou"
\item \code{param} named list of parameters for the distribution matching the input in \code{rdistr} for a given "distr".
}}
}
