% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{sde_merton}
\alias{sde_merton}
\title{Simulate a Merton jump diffusion}
\usage{
sde_merton(x0, tt, param, n = 1000)
}
\arguments{
\item{x0}{initial value}

\item{tt}{time horizon}

\item{param}{vector of mean drift, volatility,
mean-rate of jumps, mean jump size and jump-size sd.}

\item{n}{number of time sub-intervals}
}
\value{
data.frame numeric
}
\description{
{Simulate a Merton jump diffusion sample path using
a numerical stochastic integrator.}
}
