% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MilsteinMethod.R
\name{sde_milstein}
\alias{sde_milstein}
\title{Milstein method for solving autonomous SDEs}
\usage{
sde_milstein(x0, tt, drift, diffusion, diff1 = NULL, n = 1000)
}
\arguments{
\item{x0}{the initial point of the process}

\item{tt}{the time horizon to solve over}

\item{drift}{the infinitesimal drift coefficient, a single-variable function of space}

\item{diffusion}{the infinitesimal diffusion coefficient, a single-variable function of space}

\item{diff1}{the derivative of the diffusion coefficient function, can be \code{NULL} for numerical approximations.}

\item{n}{number of sub-intervals to use in time-interval partition}
}
\value{
data.frame
}
\description{
{A straight forward implementation of the Milstein method for solving SDEs via
stochastic integration. The Milstein scheme has both weak and strong order of convergence on
the order of the time-step, which is superior to the Euler-Maruyama scheme's square root of time order for
strong convergence.}
}
\details{
{The derivative of the diffusivity coefficient function may be
\code{NULL} and approximated numerically.}
}
