% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{sde_mixture}
\alias{sde_mixture}
\title{Simulate a log-normal mixture diffusion}
\usage{
sde_mixture(x0, tt, param, n = 1000, method = "em")
}
\arguments{
\item{x0}{initial value}

\item{tt}{time horizon}

\item{param}{matrix whose first row is probabilities, second drifts, third volatilities.}

\item{n}{number of time sub-intervals}

\item{method}{"em" or "rk2"}
}
\value{
data.frame numeric
}
\description{
{Simulate a log-normal mixture diffusion process sample path using
a numerical stochastic integrator.}
}
