% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{sde_ou}
\alias{sde_ou}
\title{Simulate an OU process}
\usage{
sde_ou(x0, tt, param, n = 1000, method = "em")
}
\arguments{
\item{x0}{initial value}

\item{tt}{time horizon}

\item{param}{vector of mean-reversion rate, level, and vol-of-vol}

\item{n}{number of time sub-intervals}

\item{method}{"em" or "rk2"}
}
\value{
data.frame numeric
}
\description{
{Simulate a OU process sample path using
a numerical stochastic integrator.}
}
