% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RungeKutta2.R
\name{sde_rk2}
\alias{sde_rk2}
\title{Runge-Kutta solver for Ito SDEs}
\usage{
sde_rk2(x0, tt, drift, diffusion, n = 1000)
}
\arguments{
\item{x0}{initial value of the process}

\item{tt}{length to simulate over}

\item{drift}{the drift coefficient function of space-time in the order \code{(t, x)}}

\item{diffusion}{the diffusion coefficient function of space-time, in the order \code{(t, x)}}

\item{n}{number of sub-intervals in time-discretization}
}
\value{
data.frame
}
\description{
{The stochastic Runge-Kutta for SDEs. See wikipedia for mathematical details.}
}
