% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EulerMaruyama.R
\name{sde_system}
\alias{sde_system}
\title{Solve arbitrary first order SDE systems}
\usage{
sde_system(
  IC,
  t0 = 0,
  tn = 1,
  drifts,
  diffusions,
  independent = FALSE,
  n = 1000
)
}
\arguments{
\item{IC}{a list of variables with the same names as the non-time input as drifts, in the same order with initial values}

\item{t0}{initial time}

\item{tn}{ending time}

\item{drifts}{a list of drift functions with as many variables as IC, and in the same order with the names}

\item{diffusions}{a list of volatility functions with as many variables as IC, and in the same order with the names}

\item{independent}{boolean to simulate processes driven by one Brownian motion (false) or as many independent Brownian motions as there are state variables}

\item{n}{number of sub-intervals in time-grid}
}
\value{
data.frame
}
\description{
{A general Eule-Maruyama scheme implementation for arbitrary first-order SDE systems of finite dimension.}
}
\details{
{The list of functions must have syntax \code{h(t,x,y,z)} for each element matching the elements of \code{IC} as \code{list(x = y0, y = y0, z = z0)} for example.}
}
