% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{sde_unijumps}
\alias{sde_unijumps}
\title{Simulate a uniform jump diffusion}
\usage{
sde_unijumps(x0, tt, param, n = 1000)
}
\arguments{
\item{x0}{initial value}

\item{tt}{time horizon}

\item{param}{vector of \code{(mu, volat, lambda, a, b)}}

\item{n}{number of time sub-intervals}
}
\value{
data.frame numeric
}
\description{
{Simulate a displaced uniform jump diffusion sample path using
a numerical stochastic integrator.}
}
