% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelWrappers.R
\name{sdeSolve}
\alias{sdeSolve}
\title{Numerically solve an SDE given a model/parameter set}
\usage{
sdeSolve(spot, maturity, model, n, method = "em")
}
\arguments{
\item{spot}{initial price}

\item{maturity}{time horizon to simulate over}

\item{model}{a list defining the model, see details}

\item{n}{number of variates to use}

\item{method}{numerical scheme either "em" or "rk2", only applicable to pure diffusions.}
}
\value{
data.frame of time and price and possibly volatility for two-factor models
}
\description{
{Numerically solve an SDE and generate a sample path under a variety of models:
"gbm", "mixture", "merton", "heston", for example. See details for further information.}
}
\details{
{The argument \code{model} must be a named list with elements \code{name} which is
eitehr "gbm", "mixture", "merton", "kou", "dkou", or "heston", and a vector
\code{param} containing the numeric parameters defining the aforementioned model.
\itemize{
\item for "gbm", \code{param} should be a vector of \eqn{(\mu, \sigma)}
\item for "mixture", \code{param} should be a \eqn{3 x d} matrix }
}
}
