% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePathWrappers.R
\name{samplePathIto}
\alias{samplePathIto}
\title{Simulate sample path of Ito-process}
\usage{
samplePathIto(
  region,
  dynamics = NULL,
  control = list(n = 1000, method = "em", engine = "cpp")
)
}
\arguments{
\item{region}{initial point and length of time as a vector}

\item{dynamics}{list of drift and volatility functions of \code{(t, x)}}

\item{control}{list of variates \code{n}, method and engine}
}
\value{
data.frame of time variable and space variable
}
\description{
{Simulate a sample path of a solution to an SDE.
Wrapper to internal solvers.}
}
\details{
{The argument \code{control} must be a named list of
\itemize{
\item \code{n} number of variates in simulation, default 1000}
\item \code{method} either "em" for Euler-Maruyama or "rk2" for Runge-Kutta order 2
\item \code{engine} either "r" or "cpp"}
}
