% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EulerMaruyama.R
\name{sde_em}
\alias{sde_em}
\title{Euler-Maruyama solver for Ito SDEs}
\usage{
sde_em(x0, tt, drift, diffusion, n = 1000)
}
\arguments{
\item{x0}{initial value of the process}

\item{tt}{length to simulate over}

\item{drift}{the drift coefficient function of space-time in the order \code{(t, x)}}

\item{diffusion}{the diffusion coefficient function of space-time, in the order \code{(t, x)}}

\item{n}{number of sub-intervals in time-discretization}
}
\value{
data.frame
}
\description{
{The stochastic Euler-Maruyama solver for SDEs. See wikipedia for details}
}
